Move Route into Models namespace
