+++ packages/framework/src/Contracts/AbstractPage.php
@@ -4,7 +4,7 @@ namespace Hyde\Framework\Contracts;
 
 use Hyde\Framework\Concerns\CanBeInNavigation;
 use Hyde\Framework\Concerns\HasPageMetadata;
-use Hyde\Framework\Modules\Routing\Route;
+use Hyde\Framework\Models\Route;
 use Hyde\Framework\Services\CollectionService;
 use Illuminate\Support\Collection;
