+++ packages/framework/src/Modules/Routing/Router.php
@@ -10,6 +10,7 @@ use Hyde\Framework\Models\Pages\BladePage;
 use Hyde\Framework\Models\Pages\DocumentationPage;
 use Hyde\Framework\Models\Pages\MarkdownPage;
 use Hyde\Framework\Models\Pages\MarkdownPost;
+use Hyde\Framework\Models\Route;
 use Illuminate\Support\Collection;
 
 /**
